ml-model-deployment-1/requirements.txt
tensorflow
flask
pillow
gunicorn
google-cloud-storage
pyjwt
shortuuid